Add FT support and create more files for easier management
